+++ modules/Discipline/StudentFieldBreakdown.php
@@ -104,13 +104,13 @@ if ( ! empty( $_REQUEST['field_id'] ) )
 		if ( $fields_RET[1]['TYPE'] === 'autos' )
 		{
 			// add values found in current year
-			$options_RET = DBGet( "SELECT DISTINCT " . $field_column . ",upper(" . $field_column . ") AS KEY
+			$options_RET = DBGet( "SELECT DISTINCT " . $field_column . ",upper(" . $field_column . ") AS SORT_KEY
 				FROM students s,student_enrollment sse
 				WHERE sse.STUDENT_ID=s.STUDENT_ID
 				AND (sse.SYEAR='" . UserSyear() . "')
 				AND " . $field_column . " IS NOT NULL
 				AND " . $field_column . " != ''
-				ORDER BY KEY" );
+				ORDER BY SORT_KEY" );
 
 			foreach ( (array) $options_RET as $option )
 			{